The 14.02.2020 Night
Added changes to Vision due to the change of the illumination device used when operating the automated command.
In addition, a change was introduced in the speed of the shooter, and a delay of the shooting.
i.e the shooter has time to gain speed before the loading and delivery motors start working.