Changes to autos. And a new systems check auto command.
Changes were made to the type of the neutral mode of the motors in the roulette.
In addition, new conditions were made in the isFinished() functions in the Roulette close function.

Changes were also introduced to the visionCommand class which include but are not limited to the way it works.

In addition, a new command, called GyroTurn and GyroReset, GyroReset2, CloseElevatorForSeconds, OpenElevatorForSeconds, GyroTest were added as well.

Moreover, navx guidance systems were added to the robot and to the code as well, to allow a better tracking of the robot's angle.